优化代码

+++ core/types.go
@@ -41,7 +41,7 @@ type Engine interface {
 
 	Select(v interface{}) error
 
-	Count(v interface{}) (int, error)
+	Count(v interface{}) (int64, error)
 
 	Create(v interface{}) error
